Smoother resizing

+++ DesktopClock/Properties/Settings.cs
@@ -38,7 +38,7 @@ public sealed class Settings : INotifyPropertyChanged
     public WindowPlacement Placement { get; set; }
     public bool Topmost { get; set; } = true;
     public bool ShowInTaskbar { get; set; } = true;
-    public int Height { get; set; } = 48;
+    public double Height { get; set; } = 48;
     public string TimeZone { get; set; } = string.Empty;
     public string Format { get; set; } = "dddd, MMM dd, HH:mm:ss";
     public bool BackgroundEnabled { get; set; } = true;
